Add collecting_results phase to jobs
This adds a pre-finished phase where artifacts are still processing. If you mutate the job and suggest its finished while artifacts are pending it will automatically rotate to 'collecting_results'.

Additionally this adds various SELECT FOR UPDATE locks as a default on mutation of endpoints, as well as in the aggregation tasks.