pubsub: Add draft implementation
This adds a pubsub mechanism using redis and asyncio. It restricts data access via a signed token which contains the repository IDs the tenant has access to. This token is returned with every API response as 'X-Stream-Token', and can be passed as the 'token' query parameter to the stream service.

Refs GH-8