Add draft implementation of stream consumer
This is based on the redux-websocket implementation (no idea what im doing). It uses a middleware right now to manage the state, and as a quick hack, it also automatically [re-]connects to the stream when the token is updated. The token itself is bound on the first auth request.

This would likely need some cleanup around how the middleware functions, needs CORS handling, and also needs to handle expired tokens (both on the server and consumer level). It also needs to filter out data on the UI, and likely change what its subscribed to based on where the user is.